Added basic server for communicating with database
Initial version of a server which can communicate with MySQL database. There is a bash script "create-pool.sh" which can be used in order to create a file for pool connection with MySQL server. Executing "index.js" starts the server (you can use "npm start" if you have "nodemon" installed).